Prevent NoMethodError when setting a message expectation on an instance of a package-protected class
An Java object can be an instance of a class that is
package-protected but implements or inherits from
a public type.  In that situation, the object's proxy
will only implement methods of its public interfaces
and superclasses, but `java_instance_methods` will
include methods that are defined on the
package-protected class.

For example:

    // Java code

    public interface A {
      public void runA();
    }

    class B implements A {
      public void runA() {}
      public void runB() {}
    }

    # Ruby code
    b = B.new
    b.java_class.java_instance_methods.map(&:name) #=> contains `:run_a` and `:run_b`
    b.method(:run_a) # => Returns a Method
    b.method(:run_b) # => raises a NoMethodError